рефакторинг по рекомендациям наставника
