+++ src/components/error-message/error-message.css
@@ -5,7 +5,7 @@
   z-index: 9999;
   padding: 1rem 1.5rem;
   background-color: #d96666;
-  color: #ffffff;
+  color: #fff;
   border-radius: 0.5rem;
   box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
